+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -525,6 +525,16 @@ pub(crate) fn migrate_eslint_any_rule(
                 .get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
+        "jsx-a11y/autocomplete-valid" => {
+            if !options.include_nursery {
+                return false;
+            }
+            let group = rules.nursery.get_or_insert_with(Default::default);
+            let rule = group
+                .use_valid_autocomplete
+                .get_or_insert(Default::default());
+            rule.set_level(rule_severity.into());
+        }
         "jsx-a11y/click-events-have-key-events" => {
             let group = rules.a11y.get_or_insert_with(Default::default);
             let rule = group